requirements.txt
Django==4.2.16
psycopg2-binary==2.9.9
PyNFe==0.6.0
django-ledger==0.7.10
python-decouple==3.8
gunicorn==21.2.0
dj-database-url==2.1.0